move name service to statedb
