+++ chain/db/rollback.go
@@ -45,10 +45,6 @@ func (cs *ChainStore) Rollback(b *block.Block) error {
 		return err
 	}
 
-	if err := cs.rollbackNames(b); err != nil {
-		return err
-	}
-
 	if err := cs.rollbackPubSub(b); err != nil {
 		return err
 	}
@@ -101,40 +97,6 @@ func (cs *ChainStore) rollbackCurrentBlockHash(b *block.Block) error {
 	}
 
 	return cs.st.BatchPut([]byte{byte(SYS_CurrentBlock)}, value.Bytes())
-}
-
-func (cs *ChainStore) rollbackNames(b *block.Block) error {
-	for _, txn := range b.Transactions {
-		if txn.UnsignedTx.Payload.Type == pb.REGISTER_NAME_TYPE {
-			pl, err := transaction.Unpack(txn.UnsignedTx.Payload)
-			if err != nil {
-				return err
-			}
-
-			registerNamePayload := pl.(*pb.RegisterName)
-			err = cs.DeleteName(registerNamePayload.Registrant)
-			if err != nil {
-				return err
-			}
-		}
-	}
-
-	for _, txn := range b.Transactions {
-		if txn.UnsignedTx.Payload.Type == pb.DELETE_NAME_TYPE {
-			pl, err := transaction.Unpack(txn.UnsignedTx.Payload)
-			if err != nil {
-				return err
-			}
-
-			deleteNamePayload := pl.(*pb.DeleteName)
-			err = cs.SaveName(deleteNamePayload.Registrant, deleteNamePayload.Name)
-			if err != nil {
-				return err
-			}
-		}
-	}
-
-	return nil
 }
 
 func (cs *ChainStore) rollbackPubSub(b *block.Block) error {
